Use single point of reference for sidechain atom names (SC_DATA)
Modified data_aq to use the atom names specified in Sidechains.py for
downloading. This way, there can't be errors with atom naming between
the two files. Note: there is a slight difference between the two files:
data_aq requires the C, CA atoms to build out the dihedrals, but
Sidechains.py only needs to know which atoms are predicted by the model
and which are not.