Parsing of ProteinNet dictionaries. Skips incomplete/ASTRAL proteins.
File maintains the organization of ProteinNet as much as possible. For
example, data["validation"] = {split: split_data}. Test structures are
taken from a local directory, which has been downloaded from the CASP
webserver and contains PDB files named as targets. Many simplifications
have also been made with respect to the previous `data_aq` file. IDs are
always stored as they are in ProteinNet, even if this adds an additional
step to download the PDB structures later.

Lastly, there is still an issue where the length of the mask is longer
than that of the sequence. This means there is a not-understood behavior
that results in a difference between ProteinNet's sequences + masks and
the sequences that are obtained from the PDB. It seems that most of the
masks were just padded at the start, and removing these could fix the
errors.