Add support for training with structure data.
Protein structures are recorded as (3bb + 10sc) * 3 tensors where 10
atoms are reserved for the heavy-atom sidechain structure. The atom
ordering is as defined in Sidechains module. Angles are still included
for computing RMSE loss. The drmsd will be computed between coord
tensors instead of requiring to build the coordinates from the true
angles.