Load and provide coordinate data, without using it. Coords are 13xL.
This commit is in preparation for making the loss function use
previously measured coordinate tensors instead of regenerating them from
angles. It changes the coordinate generation functions to default to
producing 13xL coordinate tensors for each protein (which is easy to
understand and parse). It's filled with 0s where there are missing
atoms.