Utilize provided instead of generated coordinates for drmsd loss.
There is likely an issue with the provided data but, assuming the
data is correct, the downstream analysis should be correct. Also makes
some assumptions, such as the size of the coordinate array. As a band-
aid, will copy zeros from the provided tensor to the target tensor.
This will all need to be revamped when padding is changed from zeros
to infs or nans. It will also be helpful to compare the drmsd calculated
from angles and from coordinates.