Add PDB_Creator object to write PDB files given coords and mappings.
The general idea is that if any model is capable of predicting a set of
coordinates and mapping between those coordinates and residue/atom
names, then this object can be use to transform that output into a
PDB file. This script does not play nicely with CUDA and will have to
be improved.