Fix bug with parallel drmsd. Must return gradients in child process. Refactor drmsd.
When the loss is computed in parallel over the batch, new processes are
made. Each process starts with a new graph. In this case, the input
is the angle tensor, and this tensor is now considered a leaf node.
We can take the angles, turn them into coordinates, compute drmsd,
and THEN when we call backward (still in the process), the input to the
parallelized loss function will have its .grad value populated. We can
then return this gradient value, along with the loss values (as scalars).
Then, we stack all of the gradients returned by the parallelized fns,
and use these to continue the backpropagation step from the angles
back through the rest of the model.