Implement a binned dataset for training with dynamic batch sizes.
The biggest effect of this commit is randomly selected batches with
a remarkable speedup in training time.

This commit now assumes that datasets are sorted in ascending order by
default. Thus, sorting arguments are deprecated.

Also, the default behavior during training is to dynamically adjust the
size of the batch to match a consistent number of residues in each batch
(the original batch size * MAX_SEQ_LEN, ie 8 * 500). Dataset bins are
computed at the start of every training run, so this could potentially
be avoided by precomputing them. The evaluation of the training set
will have to change in a future commit so that the non-binned version
may be evaluated.