Implemented a Transformer model that embeds, convolves, then attends.
This is a modest architecture improvement that uses an Encoder only
Transformer. The only change is that after the sequence is embedded
from its one-hot representation into the "model dimension", it undergoes
a series of length-preserving sequence convolutions. Each convolution
can have its own (odd-length) kernel size as well as amount of channel
dimensionality reduction.